image/requirements.txt
fastapi
python-jose
passlib[bcrypt]
requests
mangum
pydantic
starlette
uvicorn
boto3
python-multipart